Generalise relational reconciliation w/ higher order
- use of a base function for reconciling the differences between LEFT
and RIGHT, taking a CTE `function` as a string, returning the full set
based on the simple equality predicate from the CTE targeting attrbute
`id text`.

The idea is to establish an identifier that fits the semantics for the
given datatype, encode it as `text` in the query for the CTE (think of
it as returning a results set, as opposed to the inlining of the
sql code).

The function returning a set could be scary if there are large sets are
involved, as there is no possibility for the planner to fold the CTE
query into something more efficient with just the raw data, but in
these cases, n < 10000 is acceptable.